requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-chess>=1.999
websockets>=12.0
python-multipart>=0.0.6
